Moved health bar and ammo files to game folder. Fixed missing import for enemyplane file.
