+++ fort_macarthur/lib/game/game_loop.dart
@@ -9,8 +9,6 @@ import 'enemyplane.dart';
 import 'package:flame/game.dart';
 import 'package:flame/gestures.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:fort_macarthur/ammo.dart';
-import 'package:fort_macarthur/healthbar.dart';
 import 'missile_system.dart';
 
 // main game loop. pan detector necessary for touch detection
